Keep track of deviances
Determine deviance (standard deviation normalized against mean) of
numbers and highlight the ones that deviate too much from the mean.

By default, the maximum deviance accepted without warnings is 5%.
This parameter can be tweaked via the --max-deviance option for
run.scm.

This commit changes the format of the log file.