Track maximum live heap
Extract information of maximum live heap from the output of `time'.
The maximum live heap values are not printed to the standard output in
the execution of run.scm due to screen space constraints.  It only
goes to logs.

compare.scm renders the maximum live heap values in the output it
produces.

Adding maximum live heap to run.scm logs requires bumping the log
format.